requirements.txt

Flask==0.10.1
Flask-Cors==2.1.0
Flask-Jsonpify==1.5.0
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-RESTful==0.3.6
Flask-SQLAlchemy==2.0
gunicorn==19.7.1
Jinja2==2.8
MarkupSafe==0.23
Werkzeug==0.11.2
psycopg2==2.6.2
